Add modes and failsafe buttons (#33)
* Add a mode for PID driving

This adds a mode for PID driving. This also adds the driver and gunner mode to the Shuffleboard display.

* Add modes

This adds all the gunner modes.

The mode manager supports a dedicated failsafe mode. The mode manager class hierarchy was expanded to simplify usage.

Failsafe buttons were added to the Shuffleboard test tab.

* Simplify dpadmanager

This simplifies the dpad manager state, depending more on the base
manager class state.